test: align SSO/OpenID spec layouts and naming
- Drop the wizard-mode tests from `sso.spec.ts`. The file now mirrors
  `openid.spec.ts` exactly: 1-click only, single top-level describe,
  child describes named `with X` / `without X`, no mode parentheticals.
- Move the SSO-label-discovery scenario (the only thing the wizard
  tests covered uniquely) to `consent.spec.ts` as `with SSO-sourced
  attributes`, so consent UI behaviours all live in the same file.
- Rename top-level describes to `Authorize with 1-click OpenID` /
  `Authorize with 1-click SSO` (`openid-legacy.spec.ts` keeps its
  `(legacy attributes)` suffix). Both URL-param flows are explicitly
  1-click; the names now say so.
- Wrap the OpenID sign-up test in `index.spec.ts` in a `Sign up with
  OpenID` describe so it's structurally identical to the existing
  `Sign up with SSO` describe; drop the dead manual fetch.
